backend/server/requirements.txt
asgiref
Django==5.1.6
django-cors-headers
djangorestframework
pytz
sqlparse
psycopg2-binary
python-dotenv
django-cors-headers
